Refacto GeoJsonHelper and reduce functions visibilities
