+++ test/sort.spec.js
@@ -1,6 +1,6 @@
 import test from 'ava'
 import Point from '../src/Point.js'
-import { sortPoints } from '../src/createGraphFromGeoJson.js'
+import GeoJsonHelper from '../src/GeoJsonHelper.js'
 
 // sort the vertices of the obstacle polygons according to the
 // counter clock wise angle the half line from v to each vertex makes
@@ -14,7 +14,7 @@ test('Edge test', t => {
 
   const points = [p1, p2, p3, p4]
   const clonedPoints = points.slice(0)
-  sortPoints(p1, clonedPoints)
+  GeoJsonHelper.sortPoints(p1, clonedPoints)
   // expected is counter-clockwise so p4 is the first point then p2, p3
   t.deepEqual([p1, p4, p2, p3], clonedPoints)
 })
